requirements.txt
spotipy
python-dotenv
pandas
python-telegram-bot==13.15
matplotlib
jupyterthemes
beautifulsoup4